project without scoreboard
